+++ Tests/Zoo.test.ps1
@@ -37,18 +37,6 @@ task Validate {
 	Remove-Item z.psd1, $xmlPath
 }
 
-task BadXml {
-	#! on problems, the target file should not be touched
-	Remove-Item [z].psd1
-	$xml = [xml]'<bar><bad/></bar>'
-	($r = try {Export-PsdXml z.psd1 $xml} catch {$_})
-	equals "$r" "Unexpected XML element 'bad'."
-	assert (!(Test-Path z.psd1))
-
-	($r = try {Convert-XmlToPsd $xml} catch {$_})
-	equals "$r" "Unexpected XML element 'bad'."
-}
-
 task ExtractPart {
 	$xml = {
 		@{